added a fetchNativeBalance method to alchemy-sdk.js, edited the fetchNativeBalance method in DashboardIntro.jsx, made some changes to MyNfts.jsx, NftLoansModal.jsx and RepayLoansModal.jsx
